# swagger.yaml
swagger: "2.0"
info:
  title: ADAPTT API
  description: API for the ADAPTT (Transparency in Infrastructure Projects) backend.
  version: "1.0.0"
host: localhost:5001
schemes:
  - http
basePath: /
tags:
  - name: Projects
    description: Operations related to infrastructure projects
  - name: Locations
    description: Operations related to regions and locations
  - name: Users
    description: User registration and management
  - name: Subscriptions
    description: Project subscription management
  - name: Messaging
    description: Bulk messaging operations
  - name: Webhooks
    description: Twilio webhooks

paths:
  /api/projects:
    get:
      tags:
        - Projects
      summary: Get all projects
      description: Returns a list of all projects with their transparency scores.
      responses:
        200:
          description: A list of projects
          schema:
            type: array
            items:
              $ref: '#/definitions/ProjectSummary'

  /api/projects/{project_id}:
    get:
      tags:
        - Projects
      summary: Get project details
      description: Returns detailed information about a specific project.
      parameters:
        - name: project_id
          in: path
          description: ID of the project to fetch
          required: true
          type: string
      responses:
        200:
          description: Project details
          schema:
            $ref: '#/definitions/ProjectDetail'
        404:
          description: Project not found

  /api/projects/{project_id}/documents:
    get:
      tags:
        - Projects
      summary: Get project documents
      description: Returns a list of documents associated with a project.
      parameters:
        - name: project_id
          in: path
          description: ID of the project
          required: true
          type: string
      responses:
        200:
          description: List of documents
          schema:
            type: array
            items:
              $ref: '#/definitions/Document'

  /api/locations:
    get:
      tags:
        - Locations
      summary: Get all locations
      description: Returns a list of all available regions/locations.
      responses:
        200:
          description: List of locations
          schema:
            type: array
            items:
              $ref: '#/definitions/Location'

  /api/users/register:
    post:
      tags:
        - Users
      summary: Register a new user
      description: Registers a user with their phone number and region.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - name
              - phone_number
              - region_id
            properties:
              name:
                type: string
                description: User's full name
              phone_number:
                type: string
                description: Mozambican phone number (+258 XX XXX XXXX or 8X/9X XXXXXXX)
              region_id:
                type: string
                description: Region ID from locations table
      responses:
        201:
          description: User registered successfully
        400:
          description: Validation error

  /api/subscriptions:
    post:
      tags:
        - Subscriptions
      summary: Subscribe to a project
      description: Subscribes a user to receive updates for a specific project.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - user_id
              - project_id
            properties:
              user_id:
                type: integer
                description: ID of the user
              project_id:
                type: string
                description: ID of the project
              notification_channel:
                type: string
                enum: [sms, wpp]
                default: sms
                description: Preferred notification channel
      responses:
        201:
          description: Subscription created successfully
        400:
          description: Error creating subscription

    delete:
      tags:
        - Subscriptions
      summary: Unsubscribe from a project
      description: Removes a user's subscription to a project.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - user_id
              - project_id
            properties:
              user_id:
                type: integer
                description: ID of the user
              project_id:
                type: string
                description: ID of the project
      responses:
        200:
          description: Unsubscribed successfully
        400:
          description: Error unsubscribing

  /api/subscriptions/user/{user_id}:
    get:
      tags:
        - Subscriptions
      summary: Get user subscriptions
      description: Returns all projects a user is subscribed to.
      parameters:
        - name: user_id
          in: path
          description: ID of the user
          required: true
          type: integer
      responses:
        200:
          description: List of subscriptions
          schema:
            type: array
            items:
              $ref: '#/definitions/Subscription'

  /api/messages/send-bulk:
    post:
      tags:
        - Messaging
      summary: Send bulk messages
      description: Sends SMS or WhatsApp messages to a list of phone numbers.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - message
              - phone_numbers
            properties:
              message:
                type: string
                description: Message content
              phone_numbers:
                type: array
                items:
                  type: string
                description: List of phone numbers
      responses:
        200:
          description: Bulk send results
          schema:
            type: object
            properties:
              total:
                type: integer
              successful:
                type: integer
              failed:
                type: integer
              results:
                type: array
                items:
                  type: object

  /webhook/sms:
    post:
      tags:
        - Webhooks
      summary: Twilio SMS Webhook
      description: Endpoint for handling incoming SMS from Twilio.
      consumes:
        - application/x-www-form-urlencoded
      parameters:
        - name: From
          in: formData
          type: string
          description: Sender phone number
        - name: Body
          in: formData
          type: string
          description: Message body
      responses:
        200:
          description: TwiML response
          schema:
            type: string

  /webhook/whatsapp:
    post:
      tags:
        - Webhooks
      summary: Twilio WhatsApp Webhook
      description: Endpoint for handling incoming WhatsApp messages from Twilio.
      consumes:
        - application/x-www-form-urlencoded
      parameters:
        - name: From
          in: formData
          type: string
          description: Sender phone number (whatsapp:+...)
        - name: Body
          in: formData
          type: string
          description: Message body
      responses:
        200:
          description: TwiML response
          schema:
            type: string

definitions:
  ProjectSummary:
    type: object
    properties:
      project_id:
        type: string
      project_name:
        type: string
      status:
        type: string
      transparency_score:
        type: integer
      alert_color:
        type: string
      last_sync:
        type: string

  ProjectDetail:
    allOf:
      - $ref: '#/definitions/ProjectSummary'
      - type: object
        properties:
          simple_message:
            type: string
          data_raw:
            type: object
          is_processed:
            type: boolean

  Document:
    type: object
    properties:
      document_id:
        type: string
      project_id:
        type: string
      document_type:
        type: string
      title:
        type: string
      url:
        type: string
      date_published:
        type: string
      status:
        type: string

  Location:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      region:
        type: string
      country:
        type: string

  Subscription:
    type: object
    properties:
      subscription_id:
        type: integer
      user_id:
        type: integer
      project_id:
        type: string
      subscribed_at:
        type: string
      notification_enabled:
        type: integer
      notification_channel:
        type: string
      project_name:
        type: string
      status:
        type: string
      transparency_score:
        type: integer
      alert_color:
        type: string